ESP32CAM_HOG_OPENCV/src/requirements/requirements.txt
joblib==1.3.2
numpy==1.26.2
opencv-contrib-python==4.9.0.80
pandas==2.1.4
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
requests==2.31.0
